# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Lumii salon management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products/{id}/movements:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: deleteProductMovements
      tags: [stock]
      summary: >-
        Admin only. Delete a product's history (manual movements and its lines of
        counter sales), before deleting a product created by mistake. Movements of
        appointments stay with their appointments.
      responses:
        "200":
          description: Movements deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteMovementsResult"
...
  /stock-movements:
    get:
      operationId: listStockMovements
      tags: [stock]
      summary: List stock movements (newest first), optionally filtered
      parameters:
        - name: from
          in: query
          required: false
          description: "First day included, YYYY-MM-DD"
          schema:
            type: string
        - name: to
          in: query
          required: false
          description: "Last day included, YYYY-MM-DD"
          schema:
            type: string
        - name: productId
          in: query
          required: false
          schema:
            type: string
        - name: clientId
          in: query
          required: false
          schema:
            type: string
        - name: reason
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/StockMovementReason"
      responses:
        "200":
          description: List of stock movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
  /stock-movements/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: deleteStockMovement
      tags: [stock]
      summary: >-
        Admin only. Delete a movement entered by mistake and correct the stock;
        a counter sale is deleted whole. Movements of an appointment are refused (409).
      responses:
        "200":
          description: Movements deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteMovementsResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Movement of an appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /sales:
    post:
      operationId: createSale
      tags: [stock]
      summary: Record an over-the-counter sale (one movement per item) and deduct stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleInput"
      responses:
        "201":
          description: Movements created for the sale
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /sales/{saleId}/cancel:
    parameters:
      - name: saleId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: cancelSale
      tags: [stock]
      summary: Cancel an over-the-counter sale with reversing movements (stock goes back)
      responses:
        "200":
          description: Reversing movements created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Sale already cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    StockMovementReason:
      type: string
      description: >-
        Causale. vendita and uso_servizio are generated by sales and completed
        appointments; the others are manual stock changes.
      enum:
        - vendita
        - uso_servizio
        - rifornimento
        - giacenza_iniziale
        - reso
        - rettifica
        - danneggiato
        - altro
...
    DeleteMovementsResult:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: integer
          description: How many movements were removed (a whole sale counts all its lines)
    StockUnit:
      type: string
      enum: [pz, g, ml]
    StockMovement:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        productName:
          type: string
          description: Product name at the time of the movement
        productBrand:
          type: string
        reason:
          $ref: "#/components/schemas/StockMovementReason"
        quantity:
          type: number
          description: Signed change - negative leaves the stock (scarico), positive enters it (carico)
        unit:
          $ref: "#/components/schemas/StockUnit"
        unitPrice:
          type: number
          nullable: true
          description: Sale price per piece (vendita only)
        date:
          type: string
          description: "YYYY-MM-DD"
        time:
          type: string
          description: "HH:MM"
        clientId:
          type: string
          nullable: true
        clientName:
          type: string
          nullable: true
        appointmentId:
          type: string
          nullable: true
        saleId:
          type: string
          nullable: true
          description: Groups the lines of one over-the-counter sale
        note:
          type: string
          nullable: true
        userName:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - productId
        - productName
        - productBrand
        - reason
        - quantity
        - unit
        - date
        - time
        - createdAt
    SaleItemInput:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          minimum: 0
      required:
        - productId
        - quantity
        - unitPrice
    CreateSaleInput:
      type: object
      properties:
        date:
          type: string
          description: "YYYY-MM-DD"
        time:
          type: string
          description: "HH:MM"
        clientId:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SaleItemInput"
      required:
        - date
        - time
        - items
...