# lib/api-zod/src/generated/types/deleteMovementsResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lumii salon management API
 * OpenAPI spec version: 0.2.0
 */

export interface DeleteMovementsResult {
  /** How many movements were removed (a whole sale counts all its lines) */
  deleted: number;
}